move plugin into separate module
